+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- **The schema for YAML files has changed**.
+  Instead of requiring a top-level "type" key, each packet definition must contain a single-key mapping with the packet type and a mapping of its fields.
+  The "payload" key in context and command packets has also been removed; all packet fields should be under the type mapping.
+  **Existing YAML files must be updated to work with version 0.3.**
 ### Added
 - User-defined Discrete I/O fields, including indicators.
 ### Removed